priomps: split regex into separate file
